[WIP] Initial removal
