+++ bittensor_cli/src/bittensor/templates/subnet-details.j2
@@ -1,0 +1,30 @@
+<div id="subnet-modal" class="modal hidden">
+    <div class="modal-content">
+        <div class="modal-header">
+            <h2 class="subnet-title"></h2>
+            <button class="close-button" onclick="closeSubnetModal()">&times;</button>
+        </div>
+        <div class="subnet-overview">
+            <div class="overview-item">
+                <span class="label">Price</span>
+                <span class="value price"></span>
+            </div>
+            <div class="overview-item">
+                <span class="label">Market Cap</span>
+                <span class="value market-cap"></span>
+            </div>
+            <div class="overview-item">
+                <span class="label">Emission Rate</span>
+                <span class="value emission"></span>
+            </div>
+            <div class="overview-item">
+                <span class="label">Your Total Stake</span>
+                <span class="value total-stake"></span>
+            </div>
+        </div>
+        <div class="stakes-section">
+            <h3>Your Stakes</h3>
+            <div class="stakes-list"></div>
+        </div>
+    </div>
+</div>
