+++ pyproject.toml
@@ -33,7 +33,6 @@ dependencies = [
     "typer>=0.12,<0.16",
     "bittensor-wallet>=3.0.7",
     "plotille>=5.0.0",
-    "pywry>=0.6.2",
     "plotly>=6.0.0",
 ]
